refactor: extract ExplorateurIA utilities
