+++ frontend/src/pages/explorateurIA/utils.ts
@@ -1,0 +1,5 @@
+export function classNames(
+  ...values: Array<string | false | null | undefined>
+): string {
+  return values.filter(Boolean).join(" ");
+}
